Move mod/ping to module class
